Use XY Coordinates, All Tiles are Objects,
FixTile returns a field-complete tile from either a partial with short or long fields or just a string to be rendered
DefaultTile is an empty tile with some default values.
Run the tiles through this method in the gamestate file, as well as changing to column-tile instead of row-tile addressing after it occurs.
Change the Game class to use this new addressing,
Move the tile location logic to the tile's render