Move pending render to tiles, add intermediate tile render
Game no longer needs to control knowledge of if the player is pending movement because the tiles do.
Tiles know when they are targeted by pending movement by mouse enter and exit and render accordingly.
They also know when they are an intermediate step in path and render according to that.
Pathfinding has a rudimentary method for creating a path and determing if coordinates fall inside it:
Go all the way up/down and then all the way left/right.